Digital Output Setup Conditioning
Added code to initialize digital outputs used by GEVCU.  Due to a bug
in the Due version of Arduino, the digitalRead() function of an output
pin will always return zero.  The work around is to set the port as an
INPUT somewhere along the line BEFORE setting it as an output.

This addition to GEVCU setup() simply sets our eight outputs to be
INPUTS, then immediately sets them as outputs, then immediately sets
them off (LOW).

This should make getOutput() more reliable.